fix: change plugin to be a function
so each plugin options are scoped by themselves.
Meaning internal support of file based config are out.
Instead, there can be an app configurator that reads the config,
import and loads the plugin and create the app instead.

`initForTest()` is also moved out,
so that test code and live code are properly separated,
and the type variance are greatly reduced and simplified.

The standard way to expose a test plugin have not been decided yet.
Currently `@just-web/log` is exposing `logPluginForTest` named export.